[REFACTOR] BaseViewController에 공통 권한 토스트 관리 로직 통합
