+++ Atcha-iOS/Presentation/Course/CourseSearch/CourseSearchViewModel.swift
@@ -139,12 +139,12 @@ final class CourseSearchViewModel: BaseViewModel {
     // MARK: - 코스 검색 스트리밍용
     func startCourseStream() {
         
-        if isBlackoutNow() {
-            setLoading(false)
-            isServerError = true
-            courses = []
-            return
-        }
+//        if isBlackoutNow() {
+//            setLoading(false)
+//            isServerError = true
+//            courses = []
+//            return
+//        }
         
         courseStreamTask?.cancel()
         setLoading(true)
